4: Add keywords to improve block search
Sometimes a block could have aliases that help users discover it while
searching. For example, an `image` block could also want to be
discovered by `photo`. You can do so by providing an array of terms
(which can be translated). It is only allowed to add as much as three
terms per block.